extract routing code into a separate module
