+++ app/spandx.js
@@ -1,22 +1,19 @@
 #!/usr/bin/env node
 
 const http = require("http");
-const URL = require("url");
 const path = require("path");
-const fs = require("fs");
 const https = require("https");
 
 const browserSync = require("browser-sync");
 const connect = require("connect");
 const httpProxy = require("http-proxy");
 const transformerProxy = require("transformer-proxy");
-const serveStatic = require("serve-static");
-const finalhandler = require("finalhandler");
 const _ = require("lodash");
 const c = require("print-colors");
 const ESI = require("nodesi");
 const opn = require("opn");
 
+const router = require("./router.js");
 const config = require("./config");
 const resolveHome = require("./resolveHome");
 
@@ -52,17 +49,6 @@ async function init(confIn) {
     https.globalAgent.options.rejectUnauthorized = false;
 
     bs = browserSync.create();
-
-    // for each local file path in the conf, create a serveStatic object for
-    // serving that dir
-    const serveLocal = _(conf.routes)
-        .omitBy(_.isObject)
-        .mapValues(dir =>
-            serveStatic(path.resolve(conf.configDir, resolveHome(dir)), {
-                redirect: true
-            })
-        )
-        .value();
 
     const esi = _(conf.host)
         .mapValues((host, env) => {
@@ -113,87 +99,15 @@ async function init(confIn) {
     });
 
     //
-    app.use((req, res, next) => {
-        next();
-    });
+    // app.use((req, res, next) => {
+    //     next();
+    // });
 
     // apply ESI
     app.use(transformerProxy(applyESI));
 
     // dynamically proxy to local filesystem or remote webserver
-    app.use((req, res, next) => {
-        // figure out which target to proxy to based on the requested resource path
-        const routeKey = _.findKey(conf.routes, (v, r) =>
-            _.startsWith(req.url, r)
-        );
-        const env = req.headers["x-spandx-env"];
-        const route = conf.routes[routeKey];
-        let target = route.host && route.host[env];
-        // console.log(`env: ${env}`);
-        // console.log(`route: ${route}`);
-        // console.log(`target: ${target}`);
-        let fileExists;
-        let needsSlash = false;
-        const localFile = !target;
-
-        // determine if the URL path maps to a local directory
-        // if it maps to a local directory, and if the file exists, serve it
-        // up.  if the URL path maps to an HTTP server, OR if it maps to a file
-        // but the file doesn't exist, in either case proxy to a remote server.
-        if (localFile) {
-            const url = URL.parse(req.url);
-            const relativeFilePath = url.pathname.replace(
-                new RegExp(`^${routeKey}/?`),
-                "/"
-            ); // remove route path (will be replaced with disk path)
-            const absoluteFilePath = path.resolve(
-                conf.configDir,
-                resolveHome(route),
-                relativeFilePath.replace(/^\//, "")
-            );
-            fileExists = fs.existsSync(absoluteFilePath);
-
-            if (fileExists) {
-                const oldUrl = req.url;
-                const isDir = fs.lstatSync(absoluteFilePath).isDirectory();
-
-                if (conf.verbose) {
-                    console.log(
-                        `GET ${c.fg.l.green}${req.url}${c.end} from ${
-                            c.fg.l.cyan
-                        }${absoluteFilePath}${c.end} ${env}`
-                    );
-                }
-
-                req.url = relativeFilePath; // update the request's url to be relative to the on-disk dir
-                serveLocal[routeKey](req, res, finalhandler(req, res));
-                return; // stop here, don't continue to HTTP proxy section
-            }
-        }
-
-        if (localFile && (!fileExists || needsSlash)) {
-            target = conf.routes["/"].host[env];
-        }
-
-        if (conf.verbose) {
-            console.log(
-                `GET ${c.fg.l.green}${req.url}${c.end} from ${
-                    c.fg.l.blue
-                }${target}${c.end}${c.fg.l.green}${req.url}${c.end}`
-            );
-        }
-
-        if (target) {
-            proxy.web(req, res, { target }, e => {
-                console.error(e);
-                res.writeHead(200, { "Content-Type": "text/plain" });
-                res.end();
-            });
-        } else {
-            res.writeHead(404);
-            res.end();
-        }
-    });
+    app.use(router(conf, proxy));
     internalProxy = http.createServer(app).listen(internalProxyPort);
 
     // output for humans
@@ -204,41 +118,48 @@ async function init(confIn) {
         console.log("These paths will be routed to the following remote hosts");
         console.log();
         console.log(
-            _.map(conf.webRoutes, route => {
-                return conf.spandxUrl
-                    .map(
-                        url =>
-                            `  ${c.fg.l.blue}${url.replace(/\/$/, "")}${c.end}${
-                                c.fg.l.green
-                            }${route[0]}${c.e} will be routed to ${
-                                c.fg.l.blue
-                            }${route[1].host.default || route[1].host}${c.e}${
-                                c.fg.l.green
-                            }${route[0]}${c.e}`
-                    )
-                    .join("\n");
-            }).join("\n")
+            _
+                .map(conf.webRoutes, route => {
+                    return conf.spandxUrl
+                        .map(
+                            url =>
+                                `  ${c.fg.l.blue}${url.replace(/\/$/, "")}${
+                                    c.end
+                                }${c.fg.l.green}${route[0]}${
+                                    c.e
+                                } will be routed to ${c.fg.l.blue}${route[1]
+                                    .host.default || route[1].host}${c.e}${
+                                    c.fg.l.green
+                                }${route[0]}${c.e}`
+                        )
+                        .join("\n");
+                })
+                .join("\n")
         );
         console.log();
 
         console.log("These paths will be routed to your local filesystem");
         console.log();
         console.log(
-            _.map(conf.diskRoutes, route => {
-                return conf.spandxUrl
-                    .map(
-                        url =>
-                            `  ${c.fg.l.blue}${url.replace(/\/$/, "")}${c.end}${
-                                c.fg.l.green
-                            }${route[0]}${c.end} will be routed to ${
-                                c.fg.l.cyan
-                            }${path.resolve(
-                                conf.configDir,
-                                resolveHome(route[1])
-                            )}${c.e}`
-                    )
-                    .join("\n");
-            }).join("\n")
+            _
+                .map(conf.diskRoutes, route => {
+                    return conf.spandxUrl
+                        .map(
+                            url =>
+                                `  ${c.fg.l.blue}${url.replace(/\/$/, "")}${
+                                    c.end
+                                }${c.fg.l.green}${route[0]}${
+                                    c.end
+                                } will be routed to ${
+                                    c.fg.l.cyan
+                                }${path.resolve(
+                                    conf.configDir,
+                                    resolveHome(route[1])
+                                )}${c.e}`
+                        )
+                        .join("\n");
+                })
+                .join("\n")
         );
 
         console.log();
@@ -248,9 +169,9 @@ async function init(confIn) {
         );
         console.log();
         console.log(
-            _.map(conf.files, file => `  ${c.fg.l.cyan}${file}${c.e}`).join(
-                "\n"
-            )
+            _
+                .map(conf.files, file => `  ${c.fg.l.cyan}${file}${c.e}`)
+                .join("\n")
         );
         console.log();
 
@@ -259,15 +180,17 @@ async function init(confIn) {
         );
         console.log();
         console.log(
-            _.map(
-                conf.rewriteRules,
-                rule =>
-                    `  ${c.fg.l.pink}${rule.match}${
-                        c.e
-                    } will be replaced with "${c.fg.d.green}${rule.replace}${
-                        c.e
-                    }"`
-            ).join("\n")
+            _
+                .map(
+                    conf.rewriteRules,
+                    rule =>
+                        `  ${c.fg.l.pink}${rule.match}${
+                            c.e
+                        } will be replaced with "${c.fg.d.green}${
+                            rule.replace
+                        }${c.e}"`
+                )
+                .join("\n")
         );
         console.log();
     }
